Selftests: reorganize utils and plugin tests
This performs a reorganization of selftests:

 1) Moves "avocado.utils" related tests to a "utils" subdirectory
 2) Moves plugin related tests to a "plugin" subdirectory

This should give the selftests a better structure.

Also, the Avocado project has plans to split the utilities into their
own repository, and this change would enable that to be done more
easily.
