+++ selftests/functional/plugin/test_vmimage.py
@@ -5,7 +5,7 @@ import unittest.mock
 from avocado.core import exit_codes
 from avocado.utils import path, process
 
-from .. import AVOCADO, get_temporary_config
+from ... import AVOCADO, get_temporary_config
 
 
 def missing_binary(binary):
