+++ selftests/unit/plugin/test_assets.py
