+++ selftests/unit/utils/test_cloudinit.py
@@ -6,7 +6,7 @@ import unittest.mock
 
 from avocado.utils import cloudinit, data_factory, iso9660
 
-from .. import setup_avocado_loggers, temp_dir_prefix
+from ... import setup_avocado_loggers, temp_dir_prefix
 
 setup_avocado_loggers()
