+++ selftests/unit/utils/test_cpu.py
