+++ selftests/unit/utils/test_disk.py
