+++ selftests/unit/utils/test_filelock.py
@@ -4,7 +4,7 @@ import unittest
 
 from avocado.utils.filelock import AlreadyLocked, FileLock
 
-from .. import temp_dir_prefix
+from ... import temp_dir_prefix
 
 
 class TestFileLock(unittest.TestCase):
