+++ selftests/unit/utils/test_kernel.py
@@ -2,7 +2,7 @@ import unittest
 
 from avocado.utils.kernel import KernelBuild
 
-from .. import setup_avocado_loggers
+from ... import setup_avocado_loggers
 
 setup_avocado_loggers()
