+++ selftests/unit/utils/test_linux_modules.py
