+++ selftests/unit/utils/test_memory.py
