+++ selftests/unit/utils/test_process.py
@@ -7,8 +7,8 @@ import unittest.mock
 
 from avocado.utils import path, process, script
 
-from .. import (setup_avocado_loggers, skipOnLevelsInferiorThan,
-                skipUnlessPathExists)
+from ... import (setup_avocado_loggers, skipOnLevelsInferiorThan,
+                 skipUnlessPathExists)
 
 setup_avocado_loggers()
