+++ selftests/unit/utils/test_stacktrace.py
